requirements.txt
Django==1.6.5
PyChef==0.2.3
PyYAML==3.11
South==0.8.4
argparse==1.2.1
boto==2.29.1
django-allauth==0.16.1
futures==2.1.6
gunicorn==19.0.0
httplib2==0.9
jenkinsapi==0.2.19
oauthlib==0.6.2
psycopg2==2.5.3
python-dateutil==2.2
python-jenkins==0.2.1
python-memcached==1.53
python-openid==2.2.5
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
simplejson==3.5.2
six==1.7.2
wsgiref==0.1.2
